+++ packages/docs/docs/main/change-log.md
@@ -9,6 +9,11 @@ This page lists the changes to the IsaacScript framework.
 
 <br />
 
+## October 9th, 2023 (Unreleased)
+
+- Added the following helper functions:
+  - `getVanillaCollectibleTypesOfQuality`
+
 ## October 2nd, 2023
 
 - Added the following helper functions: